# docs/OpenAPI.yaml
openapi: 3.1.0
info:
    title: Course Registration System - OpenAPI 3.1
    description: |-
        This is the Course Registration System developed by Group Delta for the CMSC495 Computer Science Capstone Project.
        The API uses server-side session authentication with an HttpOnly cookie rather than JWT bearer tokens.
    license:
        name: Apache 2.0
        url: https://www.apache.org/licenses/LICENSE-2.0.html
    version: 1.0.0
...
servers:
    - url: http://127.0.0.1:3000
...
paths:
...
    /admin:
        get:
            tags:
                - admin
            summary: List users
            description: Return a paginated list of users. Requires role ADMIN. Role is verified server-side for the session user. If the authenticated user is still on first login, access is blocked until they change their password.
            operationId: getUsers
            parameters:
                - name: page
                  in: query
                  description: Page number, default = 1
                  required: false
                  schema:
                      type: integer
                - name: limit
                  in: query
                  description: Items per page, default is 10
                  required: false
                  schema:
                      type: integer
                - name: search
                  in: query
                  description: Search term
                  required: false
                  schema:
                      type: string
                - name: role
                  in: query
                  description: 'Role filter: ADMIN, STUDENT, or PROFESSOR'
                  required: false
                  schema:
                      type: string
            responses:
                '200':
                    description: Paginated user list returned successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserMeta'
                '400':
                    description: Invalid role filter
                '401':
                    description: Invalid or missing session
                '403':
                    description: 'Forbidden: ADMIN role required, or password change required on first login'
            security:
                - SessionAuth: []
        post:
            tags:
                - admin
            summary: Create a user
            description: Create a new user account. Requires role ADMIN. Role is verified server-side for the session user. If the authenticated user is still on first login, access is blocked until they change their password.
            operationId: addUser
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AdminUserCreateRequest'
            responses:
                '201':
                    description: User created successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                '400':
                    description: Invalid request body
                '401':
                    description: Invalid or missing session
                '403':
                    description: 'Forbidden: ADMIN role required, or password change required on first login'
                '409':
                    description: User with this email already exists
            security:
                - SessionAuth: []
...
    /course:
        get:
            tags:
                - course
            summary: List courses
            description: Return a paginated list of courses.
            operationId: getCourses
            parameters:
                - name: page
                  in: query
                  description: Page number, default = 1
                  required: false
                  schema:
                      type: integer
                - name: limit
                  in: query
                  description: Items per page, default is 10
                  required: false
                  schema:
                      type: integer
                - name: search
                  in: query
                  description: Search term
                  required: false
                  schema:
                      type: string
                - name: subject
                  in: query
                  description: Subject code filter
                  required: false
                  schema:
                      type: string
            responses:
                '200':
                    description: Paginated course list returned successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CourseMeta'
                '400':
                    description: Invalid query parameters
        post:
            tags:
                - course
            summary: Create a course
            description: Create a new course. Requires role ADMIN. Role is verified server-side for the session user. If the authenticated user is still on first login, access is blocked until they change their password.
            operationId: addCourse
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CourseWriteRequest'
            responses:
                '201':
                    description: Course created successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Course'
                '400':
                    description: Invalid request body
                '401':
                    description: Invalid or missing session
                '403':
                    description: 'Forbidden: ADMIN role required, or password change required on first login'
                '409':
                    description: Course with this course code already exists
                '500':
                    description: Failed to retrieve newly created course
            security:
                - SessionAuth: []
...
    /prerequisite:
        post:
            tags:
                - prerequisite
            summary: Create a prerequisite
            description: Add a prerequisite relationship to a course. Requires role ADMIN. Role is verified server-side for the session user. If the authenticated user is still on first login, access is blocked until they change their password.
            operationId: addPrerequisite
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PrerequisiteCreateRequest'
            responses:
                '201':
                    description: Prerequisite created successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Prerequisite'
                '400':
                    description: Invalid request body or a course cannot be its own prerequisite
                '401':
                    description: Invalid or missing session
                '403':
                    description: 'Forbidden: ADMIN role required, or password change required on first login'
                '404':
                    description: One or more courses not found
                '409':
                    description: Prerequisite relationship already exists or would create a cycle
                '500':
                    description: Failed to add prerequisite relationship
            security:
                - SessionAuth: []
...
    /semester:
        get:
            tags:
                - semester
            summary: List semesters
            description: Return all semesters.
            operationId: getSemesters
            parameters:
            responses:
                '200':
                    description: Semester list returned successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SemesterArray'
        post:
            tags:
                - semester
            summary: Create a semester
            description: Create a new semester. Requires role ADMIN. Role is verified server-side for the session user. If the authenticated user is still on first login, access is blocked until they change their password.
            operationId: addSemester
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SemesterWriteRequest'
            responses:
                '201':
                    description: Semester created successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Semester'
                '400':
                    description: Invalid request body
                '401':
                    description: Invalid or missing session
                '403':
                    description: 'Forbidden: ADMIN role required, or password change required on first login'
                '409':
                    description: Semester already exists
            security:
                - SessionAuth: []
...
    /section:
        get:
            tags:
                - section
            summary: List sections
            description: Return a paginated list of sections.
            operationId: getSections
            parameters:
                - name: page
                  in: query
                  description: Page number, default = 1
                  required: false
                  schema:
                      type: integer
                - name: limit
                  in: query
                  description: Items per page, default is 10
                  required: false
                  schema:
                      type: integer
                - name: search
                  in: query
                  description: Search term
                  required: false
                  schema:
                      type: string
                - name: crsId
                  in: query
                  description: Course ID, default = null
                  required: false
                  schema:
                      type: integer
                - name: semId
                  in: query
                  description: Semester ID, default = null
                  required: false
                  schema:
                      type: integer
                - name: profId
                  in: query
                  description: Professor ID, default = null
                  required: false
                  schema:
                      type: integer
                - name: subject
                  in: query
                  description: Subject code filter, for example CMSC
                  required: false
                  schema:
                      type: string
                - name: days
                  in: query
                  description: Canonical meeting days filter, for example MW or MTWR
                  required: false
                  schema:
                      type: string
            responses:
                '200':
                    description: Paginated section list returned successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ReadableSectionMeta'
                '400':
                    description: Invalid query parameters
...
    /enrollment:
        get:
            tags:
                - enrollment
            summary: List enrollments for a student
            description: Return all enrollments for a specific student. Requires role STUDENT or ADMIN. Role is verified server-side for the session user. Students may only view their own enrollments. If the authenticated user is still on first login, access is blocked until they change their password.
            operationId: getEnrollments
            parameters:
                - name: stuId
                  in: query
                  description: Student ID
                  required: true
                  schema:
                      type: integer
            responses:
                '200':
                    description: Enrollment list returned successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/EnrollmentArray'
                '400':
                    description: Invalid query parameters
                '401':
                    description: Invalid or missing session
                '403':
                    description: 'Forbidden: STUDENT or ADMIN role required, password change required on first login, and students may only access their own enrollments'
                '404':
                    description: Student not found
            security:
                - SessionAuth: []
        post:
            tags:
                - enrollment
            summary: Create an enrollment
            description: Create a new enrollment record. Requires role STUDENT or ADMIN. Role is verified server-side for the session user. Students may only manage their own enrollments and may only enroll in sections from the current semester. If the authenticated user is still on first login, access is blocked until they change their password.
            operationId: addEnrollment
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EnrollmentCreateRequest'
            responses:
                '201':
                    description: Enrollment created successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Enrollment'
                '400':
                    description: Invalid request body
                '401':
                    description: Invalid or missing session
                '403':
                    description: 'Forbidden: STUDENT or ADMIN role required, password change required on first login, students may only create their own enrollments, and students may only enroll in sections from the current semester'
                '404':
                    description: Student or section not found
                '409':
                    description: Schedule conflict, unmet prerequisites, duplicate active enrollment, or section / waitlist capacity conflict
            security:
                - SessionAuth: []
...
components:
    securitySchemes:
        SessionAuth:
            type: apiKey
            in: cookie
            name: sid
            description: HttpOnly session cookie containing a session identifier. The server resolves the user from the session and verifies role server-side for restricted operations.
    schemas:
...
        User:
            type: object
            required:
                - id
                - name
                - email
                - first_login
                - role
                - role_id
                - role_details
            properties:
                id:
                    type: integer
                    format: int32
                name:
                    type: string
                email:
                    type: string
                    format: email
                first_login:
                    type: boolean
                role:
                    type: string
                role_id:
                    type: integer
                role_details:
                    type: string
...
        AdminUserCreateRequest:
            type: object
            required:
                - name
                - email
                - detail
                - type
            properties:
                name:
                    type: string
                email:
                    type: string
                    format: email
                detail:
                    type: string
                type:
                    type: string
                    enum:
                        - ADMIN
                        - STUDENT
                        - PROFESSOR
...
        UserArray:
            type: array
            items:
                type: object
                properties:
                    User:
                        $ref: '#/components/schemas/User'
        UserMeta:
            type: object
            properties:
                User:
                    $ref: '#/components/schemas/UserArray'
                Meta:
                    $ref: '#/components/schemas/Meta'
        Meta:
            type: object
            required:
                - page
                - limit
                - total
                - totalPages
            properties:
                page:
                    type: integer
                limit:
                    type: integer
                total:
                    type: integer
                totalPages:
                    type: integer
        Course:
            type: object
            required:
                - course_id
                - course_code
                - title
                - description
                - credits
            properties:
                course_id:
                    type: integer
                course_code:
                    type: string
                title:
                    type: string
                description:
                    type: string
                credits:
                    type: integer
                subject:
                    type: string
                    description: Present on course-list responses. Single-course reads do not currently include this derived field.
        CourseWriteRequest:
            type: object
            required:
                - code
                - title
                - desc
                - cred
            properties:
                code:
                    type: string
                title:
                    type: string
                desc:
                    type: string
                cred:
                    type: integer
        CourseArray:
            type: array
            items:
                type: object
                properties:
                    Course:
                        $ref: '#/components/schemas/Course'
        CourseMeta:
            type: object
            properties:
                Course:
                    $ref: '#/components/schemas/CourseArray'
                Meta:
                    $ref: '#/components/schemas/Meta'
        Prerequisite:
            type: object
            properties:
                courseId:
                    type: integer
                courseCode:
                    type: string
                title:
                    type: string
        PrerequisiteCreateRequest:
            type: object
            required:
                - cId
                - pId
            properties:
                cId:
                    type: integer
                pId:
                    type: integer
...
        Semester:
            type: object
            required:
                - semester_id
                - term
                - year
            properties:
                semester_id:
                    type: integer
                term:
                    type: string
                year:
                    type: integer
        SemesterWriteRequest:
            type: object
            required:
                - term
                - year
            properties:
                term:
                    type: string
                year:
                    type: integer
        SemesterArray:
            type: array
            items:
                type: object
                properties:
                    Semester:
                        $ref: '#/components/schemas/Semester'
...
        SectionCourseSummary:
            type: object
            required:
                - course_id
                - course_code
                - title
                - description
                - subject
                - credits
            properties:
                course_id:
                    type: integer
                course_code:
                    type: string
                title:
                    type: string
                description:
                    type: string
                subject:
                    type: string
                credits:
                    type: integer
        SectionProfessorSummary:
            type: object
            required:
                - professor_id
                - professor_name
            properties:
                professor_id:
                    type: integer
                professor_name:
                    type: string
        ReadableSection:
            type: object
            required:
                - section_id
                - capacity
                - enrolled_count
                - waitlisted_count
                - seats_available
                - days
                - start_time
                - end_time
                - course
                - professor
                - semester
            properties:
                section_id:
                    type: integer
                capacity:
                    type: integer
                enrolled_count:
                    type: integer
                    description: Number of currently enrolled students in the section.
                waitlisted_count:
                    type: integer
                    description: Number of currently waitlisted students in the section.
                seats_available:
                    type: integer
                    description: Remaining available enrolled seats.
                days:
                    type: string
                    description: Meeting-day code string such as MWF or TR, or `async` for asynchronous classes.
                start_time:
                    type: string
                    description: Time of day in HH:MM:SS format, or an empty string for asynchronous classes.
                    pattern: '^(\d{2}:\d{2}:\d{2})?$'
                end_time:
                    type: string
                    description: Time of day in HH:MM:SS format, or an empty string for asynchronous classes.
                    pattern: '^(\d{2}:\d{2}:\d{2})?$'
                course:
                    $ref: '#/components/schemas/SectionCourseSummary'
                professor:
                    $ref: '#/components/schemas/SectionProfessorSummary'
                semester:
                    $ref: '#/components/schemas/Semester'
...
        ReadableSectionArray:
            type: array
            items:
                type: object
                properties:
                    Section:
                        $ref: '#/components/schemas/ReadableSection'
        ReadableSectionMeta:
            type: object
            properties:
                Section:
                    $ref: '#/components/schemas/ReadableSectionArray'
                Meta:
                    $ref: '#/components/schemas/Meta'
        Enrollment:
            type: object
            required:
                - enrollment_id
                - student_id
                - section_id
                - status
            properties:
                enrollment_id:
                    type: integer
                student_id:
                    type: integer
                section_id:
                    type: integer
                status:
                    type: string
                    enum:
                        - enrolled
                        - waitlisted
                        - dropped
                        - completed
        EnrollmentArray:
            type: array
            items:
                type: object
                properties:
                    Enrollment:
                        $ref: '#/components/schemas/Enrollment'
        EnrollmentCreateRequest:
            type: object
            required:
                - secId
                - stuId
            properties:
                secId:
                    type: integer
                stuId:
                    type: integer
                code:
                    type: string
                    description: Optional access code field. The current API ignores unknown create-body fields and does not use access codes during initial enrollment creation.
...